requirements.txt
airtable-python-wrapper==0.12.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
idna==2.8
py3dns==3.2.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
validate-email==1.3
whitenoise==4.1.3
